fix(nordea): Fetch interest rate from subloan when present.
When the loan is split into multiple parts the interest rate is not available in the interest attribute, but instead in the sub loan interest attribute.

+++ src/integration/agents/src/main/java/se/tink/backend/aggregation/agents/nxgen/se/banks/nordea/v30/fetcher/loan/entities/CreditEntity.java
@@ -1,23 +1,24 @@
 package se.tink.backend.aggregation.agents.nxgen.se.banks.nordea.v30.fetcher.loan.entities;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
+import java.math.BigDecimal;
 import se.tink.backend.aggregation.annotations.JsonObject;
 
 @JsonObject
 public class CreditEntity {
-    @JsonProperty private double limit;
-    @JsonProperty private double available;
-    @JsonProperty private double spent;
+    @JsonProperty private BigDecimal limit;
+    @JsonProperty private BigDecimal available;
+    @JsonProperty private BigDecimal spent;
 
-    public double getLimit() {
+    public BigDecimal getLimit() {
         return limit;
     }
 
-    public double getAvailable() {
+    public BigDecimal getAvailable() {
         return available;
     }
 
-    public double getSpent() {
+    public BigDecimal getSpent() {
         return spent;
     }
 }
